requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
firebase-admin==6.2.0
PyJWT==2.8.0
slowapi==0.1.9
python-dotenv==1.0.0
requests==2.31.0
